+++ src/containers/App.css
@@ -1,74 +1,7 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-.title{
-  color: gray;
-  font-size: 30px;
-}
-.framegent{
-  border: 1px solid gray;
-  margin: 5px auto;
-  width: 40%;
-}
-.noMap{
-  border: 1px solid black;
-  width: 20%;
-  height: 60px;
-  display: inline;
-}
 .contentAlign {
+  display:flex;
+  justify-content: space-between;
+  align-items: center;
   display: inline;
   width: 70%;
-}
-.titleh3{
-  color: black; 
-  font-size: 27px;
-  font-weight: 700;
-  text-align: center;
-}
-.website{
-  position: relative;
-  left:45%;
-}
-.lastupd{
-  position: relative;
-  left: 45%;
-  margin: 3% 0;
-  color: gray;
-
 }
