+++ src/containers/Header/Header.module.css
@@ -1,0 +1,12 @@
+.title{
+    color: gray;
+    font-size: 30px;
+}
+
+.input_text{
+    padding: 10;
+    border-radius: 5px;
+    
+
+}
+
